+++ font/type3/font.go
@@ -169,7 +169,7 @@ func (t3 *Builder) Embed(instName string) (*font.Font, error) {
 				gid := t3.cmap[r]
 				gg[i].Gid = gid
 				gg[i].Chars = []rune{r}
-				gg[i].Advance = t3.glyphWidth[gid]
+				gg[i].Advance = int32(t3.glyphWidth[gid])
 			}
 
 			return gg
